+++ libs/vfs/src/lib.rs
@@ -9,6 +9,7 @@ extern crate alloc;
 
 pub use alloc::sync::Arc;
 
+pub mod acl;
 pub mod addr;
 pub mod anon;
 pub mod cred;
@@ -39,6 +40,7 @@ pub mod superblock;
 pub mod sync;
 pub mod sysctl;
 pub mod timerfd;
+pub mod xattr;
 
 use core::sync::atomic::{AtomicU64, AtomicUsize, Ordering};
 use cred::Credentials;