+++ libs/vfs/src/tests/mod.rs
@@ -9,4 +9,5 @@ mod test_mount;
 mod test_operation;
 mod test_pipe;
 mod test_socket;
+mod test_xattr;
 mod test_stat;